requirements.txt
PyQt6
mss
keyboard
google-generativeai
Pillow
pyinstaller
python-dotenv
numpy
soundcard
